Make PieChartBase So there's less repeated code for pie charts
